requirements.txt
numpy
pandas
scikit-learn
matplotlib
seaborn
scipy
pymongo
pymongo[srv]
python-dotenv
jupyter
ipykernel
certifi
PyYAML

-e .